% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEImage.R
\name{EEImage}
\alias{EEImage}
\alias{EEImageCollection}
\title{Create \code{SpatRaster} and \code{SpatRasterCollection} from Google Earth Engine Image and ImageCollection}
\usage{
EEImage(x, silent = TRUE, ...)

EEImageCollection(
  x,
  ...,
  n = 10,
  WHERE = NULL,
  query = paste("SELECT * FROM", get_asset_ogr_table_name(x), WHERE, "LIMIT ", n),
  silent = TRUE
)
}
\arguments{
\item{x}{Path to Google Earth Engine Image/ImageCollection asset}

\item{silent}{logical. Silence warning/error messages? Default: \code{TRUE}}

\item{...}{Additional arguments to \code{terra::rast()}}

\item{n}{integer. Maximum of elements of ImageCollection to return added to the default \code{query}. Defaults to \code{10}; large collections may take a while to query.}

\item{WHERE}{character. Optional \code{WHERE} clause added to the default \code{query}.}

\item{query}{character. Query argument used by \code{terra::vect()}. Defaults to \code{paste("SELECT * FROM", get_asset_ogr_table_name(x), WHERE, "LIMIT ", n)}}
}
\value{
\code{EEImage()}: a \code{SpatRaster}

\code{EEImageCollection()}: a \code{SpatRasterCollection} where each element is a (sub)dataset from the source ImageCollection \code{x}
}
\description{
Create \code{SpatRaster} and \code{SpatRasterCollection} from Google Earth Engine Image and ImageCollection
}
\examples{
\dontrun{
EEImage("projects/earthengine-public/assets/COPERNICUS/S2/20170430T190351_20170430T190351_T10SEG")
}
\dontrun{
EEImageCollection("projects/earthengine-public/assets/COPERNICUS/S2", n = 5)
}
}
