% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{set_google_application_credentials}
\alias{set_google_application_credentials}
\title{Set Google Authentication Credentials}
\usage{
set_google_application_credentials(x)
}
\arguments{
\item{x}{Path to Google "service account" Authentication Credentials file (JSON format)}
}
\value{
Sets \code{GOOGLE_APPLICATION_CREDENTIALS} system environment variable for the current session
}
\description{
Set Google Authentication Credentials
}
\examples{
\dontrun{
set_google_application_credentials("~/example-gizmo-999999-999999999999.json")
}
}
