% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terragee-package.R
\docType{package}
\name{terragee-package}
\alias{terragee}
\alias{terragee-package}
\title{terragee: Use Google Earth Engine Assets with terra via Earth Engine Data Access GDAL Drivers}
\description{
Wrapper functions for extending and testing {terra} package with EEDA \url{https://gdal.org/drivers/vector/eeda.html} and EEDAI \url{https://gdal.org/drivers/raster/eedai.html} drivers for Google Earth Engine assets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/terragee}
  \item Report bugs at \url{https://github.com/brownag/terragee/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com}

}
\keyword{internal}
